Temporarily reverts download process and adds Mac packaging script
