+++ package.json
@@ -5,20 +5,22 @@
   "main": "index.js",
   "scripts": {
     "start": "electron .",
-    "test": "jest"
+    "test": "jest",
+    "package-mac": "electron-packager . --overwrite --platform=darwin --arch=x64 --prune=true --out=release-builds"
   },
   "author": "Darin Minamoto",
   "license": "GNU GPL 3",
   "devDependencies": {
     "babel-jest": "^18.0.0",
-    "babel-plugin-transform-object-rest-spread": "^6.20.2",
     "electron": "^1.4.13",
+    "electron-packager": "^8.5.2",
     "jest": "^18.1.0",
     "react-test-renderer": "^15.4.1"
   },
   "dependencies": {
     "babel-core": "^6.21.0",
     "babel-plugin-transform-react-jsx": "^6.8.0",
+    "babel-plugin-transform-object-rest-spread": "^6.20.2",
     "babel-preset-es2015": "^6.18.0",
     "babel-register": "^6.18.0",
     "bluebird": "^3.4.7",
@@ -41,5 +43,11 @@
     "redux": "^3.6.0",
     "redux-logger": "^2.7.4",
     "redux-thunk": "^2.1.0"
+  },
+  "jest": {
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/release-builds/"
+    ]
   }
 }
